requirements.txt
Django==2.1.2
django-cors-headers==2.4.0
django-filter==2.0.0
django-rest-marshmallow==3.1.1
djangorestframework==3.8.2
Markdown==3.0.1
marshmallow==2.15.6
pytz==2018.5
